Updated bootstrap utility and dependencies. First public pre alpha release.

+++ project/BootstrapUtility.scala
@@ -159,8 +159,10 @@ object BootstrapUtility {
         if (file.isFile) {
           file.delete()
         } else {
-          createOrEmptyFolder(file.getAbsolutePath)
-          file.delete()
+          if (file.getName != ".git") {
+            createOrEmptyFolder(file.getAbsolutePath)
+            file.delete()
+          }
         }
       }
     } else {
